Separate middleware function into middlewares folder
