+++ routes/problemslRouter.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = express.Router();
 
-const ProblemStatement = require('../models/ProblemStatements');
+const getProblem = require('../middlewares/fetchProblem');
 
 const {
   getProblems,
@@ -11,22 +11,6 @@ const {
   updateProblem,
   deleteProblem,
 } = require('../controllers/problemsController');
-
-// Middleware to get a single problem statement
-const getProblem = async (req, res, next) => {
-  try {
-    var problem = await ProblemStatement.findById(req.params.id);
-    if (problem === null)
-      return res
-        .status(404)
-        .json({ message: 'Can not find problem statement.' });
-  } catch (error) {
-    return res.status(500).json({ message: error.message });
-  }
-
-  res.problem = problem;
-  next();
-};
 
 router.get('/', getProblems);
 router.get('/add', getNewProblemForm);
